modificacao dos tipos de int para double

+++ lab01_Q1_cabecalho.hpp
@@ -4,7 +4,7 @@
 #define PI 3.1415
 using namespace std;
 
-void func_planas(int indice, int param1, int param2);
+void func_planas(double indice, double param1, double param2);
 
-void func_espaciais(int indice, int param1, int param2, int param3);
+void func_espaciais(double indice, double param1, double param2, double param3);
